requirements/base.txt
# Base
pytz==2021.1
python-slugify>=5.0.2
pillow>=7.1.0
psycopg2==2.8.5 --no-binary psycopg2

# Django
django==3.2.3
# Environment
django-environ==0.4.5
# Crum User
django-crum==0.7.6
# Inputs Forms
django-widget-tweaks==1.4.8
# Export Import
django-import-export==2.3.0
# Merge on Office Open XML
docx-mailmerge==0.5.0
# Django-Cleanup
django-cleanup==5.2.0
# Django REST Framework
#djangorestframework >= 3.9.1
#django-filter==2.0.0
#django-cors-headers >= 3.0.2

# JWT
pyjwt==1.7.1

# Passwords security
argon2-cffi==20.1.0

# Static files
#whitenoise==4.1.2

# Celery
redis>=3.2.0
django-redis==4.10.0

celery>=4.3.0
tornado>=4.2.0, <6.0.0
flower==0.9.3

